Extract common fake behaviour (FakeLocalIO and FakeSocket) into superclass FakeIO
