+++ src/test/java/com/andreamazzarella/support/FakeSocket.java
@@ -4,24 +4,12 @@ import com.andreamazzarella.chat_server.Connection;
 
 import java.io.*;
 import java.net.SocketAddress;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
-public class FakeSocket implements Connection {
+public class FakeSocket extends FakeIO implements Connection {
 
-    private OutputStream outputStream = new MonitoredByteArrayOutputStream();
-    private PipedOutputStream pipedOutputStream;
-    private PipedInputStream pipedInputStream;
-
-    private CountDownLatch waitForMessage = new CountDownLatch(1);
-
-    public FakeSocket() {
-        this.pipedOutputStream = new PipedOutputStream();
-        try {
-            this.pipedInputStream = new PipedInputStream(pipedOutputStream);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    @Override
+    public void connect(SocketAddress serverAddress) throws IOException {
+        // pretend it just made a connection (Has to be implemented as part of the Connection interface)
     }
 
     @Override
@@ -30,44 +18,7 @@ public class FakeSocket implements Connection {
     }
 
     @Override
-    public void connect(SocketAddress serverAddress) throws IOException {
-        // pretend it just made a connection
-    }
-
-    @Override
     public OutputStream getOutputStream() {
         return outputStream;
-    }
-
-    public String receivedMessages() {
-        return outputStream.toString();
-    }
-
-    public void newMessage(String message) {
-        try {
-            pipedOutputStream.write(message.getBytes());
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public void waitForMessageThen(int timeOut, TimeUnit timeUnit, Runnable callback) {
-        try {
-            waitForMessage.await();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        callback.run();
-    }
-
-    private class MonitoredByteArrayOutputStream extends ByteArrayOutputStream {
-
-        @Override
-        public synchronized void write(byte[] b, int off, int len) {
-            waitForMessage.countDown();
-            super.write(b, off, len);
-        }
-
     }
 }
